Downgrade to Poetry 1.0.10
The problem with creating virtual environments got more serious because
the setuptools package fails also to install as a transitive dependency.

Poetry 1.0.10 is still using the `venv` module to create virtual
environments, not the virtualenv package.

The `--remove-untracked` option had to be removed because it not
available for this Poetry version.